Revert "Merge remote-tracking branch 'upstream/trunk' into update/twenty-twenty-one-border-support"
This reverts commit 6e36fc008e4d430e774e33e56db33ff5b2d3b229, reversing
changes made to f21dff44af5de22f5ba983baf6264cf041f7cdec.

+++ src/wp-admin/includes/nav-menu.php
@@ -874,7 +874,7 @@ function wp_nav_menu_item_taxonomy_meta_box( $data_object, $box ) {
 		return;
 	}
 
-	$num_pages = (int) ceil(
+	$num_pages = ceil(
 		wp_count_terms(
 			array_merge(
 				$args,
